+++ persistent-postgresql/test/main.hs
@@ -49,6 +49,7 @@ import qualified MigrationSpec
 import qualified MigrationTest
 import qualified MpsCustomPrefixTest
 import qualified MpsNoPrefixTest
+import qualified NullsNotDistinctTest
 import qualified PersistUniqueTest
 import qualified PersistentTest
 import qualified PgIntervalTest
@@ -126,6 +127,7 @@ main = do
             , EmbedOrderTest.embedOrderMigrate
             , LargeNumberTest.numberMigrate
             , UniqueTest.uniqueMigrate
+            , NullsNotDistinctTest.standardUniqueMigrate
             , MaxLenTest.maxlenMigrate
             , MaybeFieldDefsTest.maybeFieldDefMigrate
             , TypeLitFieldDefsTest.typeLitFieldDefsMigrate
@@ -146,6 +148,12 @@ main = do
             , UpsertWhere.upsertWhereMigrate
             , ImplicitUuidSpec.implicitUuidMigrate
             ]
+        -- UNIQUE NULLS NOT DISTINCT is only valid on PostgreSQL 15+, so only run
+        -- that migration when the server supports it. The corresponding tests are
+        -- version-gated and mark themselves pending otherwise.
+        supportsNullsNotDistinct <- NullsNotDistinctTest.isPostgres15OrHigher
+        when supportsNullsNotDistinct $
+            setup NullsNotDistinctTest.nullsNotDistinctMigrate
         PersistentTest.cleanDB
         ForeignKey.cleanDB
 
@@ -225,4 +233,5 @@ main = do
         UpsertWhere.specs
         PgIntervalTest.specs
         ArrayAggTest.specs
+        NullsNotDistinctTest.specs
         GeneratedColumnTestSQL.specsWith runConnAssert